Version 0.2.0 (#1)
* new structure for sim compatibility

* metadata and toml update

* Left and right counter test app
