+++ metadata.json
@@ -1,0 +1,6 @@
+{
+    "callable": "BadgeBotApp",
+    "name": "BadgeBot",
+    "category": "unknown",
+    "hidden": false
+}
