+++ tildagon.toml
@@ -1,6 +1,6 @@
 [app]
 # The name of your app as displayed in the menu
-name = "BadgeBot App"
+name = "BadgeBot"
 
 # The submenu where your app should appear.
 # One of: "Badge", "Music", "Media", "Apps", "Games"
@@ -37,4 +37,4 @@ description = "A demo tildagon app for use with Hex Drive hexpansion "
 # increased, we interpret this as a new version being released.
 #
 # Version number must be an integer!
-version = "0.0.1"
+version = "0.2.0"
